# Remove chunks that do not intersect target area in gradient search resampling