--- a/pyresample/gradient/__init__.py
+++ b/pyresample/gradient/__init__.py
@@ -180,25 +180,44 @@ def reshape_to_stacked_3d(array):
     return np.stack(layers, axis=-1)
 
 
-def get_border(x_coords, y_coords):
+def get_border(x_coords, y_coords, x_stride=1, y_stride=1):
     """Get the border x- and y-coordinates."""
-    up_x = x_coords[0, :]
-    down_x = x_coords[-1, :]
-    left_x = x_coords[:, 0]
-    right_x = x_coords[:, -1]
-    up_y = y_coords[0, :]
-    down_y = y_coords[-1, :]
-    left_y = y_coords[:, 0]
-    right_y = y_coords[:, -1]
-    x_s = np.concatenate((up_x, right_x, down_x, left_x))
-    y_s = np.concatenate((up_y, right_y, down_y, left_y))
+    up_x = x_coords[0, ::x_stride]
+    right_x = x_coords[::y_stride, -1]
+    down_x = x_coords[-1, ::-x_stride]
+    left_x = x_coords[::-y_stride, 0]
+    up_y = y_coords[0, ::x_stride]
+    right_y = y_coords[::y_stride, -1]
+    down_y = y_coords[-1, ::-x_stride]
+    left_y = y_coords[::-y_stride, 0]
+    res = da.compute(up_x, right_x, down_x, left_x, up_y, right_y, down_y, left_y)
+    x_s = np.concatenate(res[0:4])
+    y_s = np.concatenate(res[4:])
+
+    return x_s, y_s
+
+
+def get_corners(x_coords, y_coords):
+    """Get the border x- and y-coordinates."""
+    x1 = x_coords[0, 0]
+    x2 = x_coords[0, -1]
+    x3 = x_coords[-1, -1]
+    x4 = x_coords[-1, 0]
+    y1 = y_coords[0, 0]
+    y2 = y_coords[0, -1]
+    y3 = y_coords[-1, -1]
+    y4 = y_coords[-1, 0]
+    res = da.compute(x1, x2, x3, x4, y1, y2, y3, y4)
+    x_s = np.array(res[0:4])
+    y_s = np.array(res[4:])
 
     return x_s, y_s
 
 
 def get_boundary(x_coords, y_coords):
     """Get boundary from 2D *x_coords* and *y_coords* arrays."""
-    x_border, y_border = get_border(x_coords, y_coords)
+    # x_border, y_border = get_border(x_coords, y_coords)
+    x_border, y_border = get_corners(x_coords, y_coords)
     boundary = [(x_border[i], y_border[i]) for i in range(len(x_border))]
 
     return boundary
@@ -218,11 +237,12 @@ def remove_extra_chunks(arrays, dst_x, dst_y):
 
     # This is the slowest bit
     # Persisting is slower
-    dst_x, dst_y = da.compute(dst_x, dst_y)
+    # dst_x, dst_y = da.compute(dst_x, dst_y)
 
     dst_polygon = get_polygon(dst_x, dst_y)
     num_chunks = arrays[0].shape[-1]
     src_x, src_y = da.compute(arrays[1], arrays[2])
+    # src_x, src_y = arrays[1], arrays[2]
     src_x = np.split(src_x, num_chunks, axis=-1)
     src_y = np.split(src_y, num_chunks, axis=-1)
     src_polys = [get_polygon(x, y) for (x, y) in zip(src_x, src_y)]
@@ -240,12 +260,14 @@ def remove_extra_chunks(arrays, dst_x, dst_y):
     # import matplotlib.pyplot as plt
     # for i in range(len(src_x)):
     #     src_xb, src_yb = get_border(src_x[i], src_y[i])
+    #     src_xb, src_yb = get_corners(src_x[i], src_y[i])
     #     if covers[i]:
     #         color = 'bx'
     #     else:
     #         color = 'r.'
     #     plt.plot(src_xb, src_yb, color)
     # dst_xb, dst_yb = get_border(dst_x, dst_y)
+    # dst_xb, dst_yb = get_corners(dst_x, dst_y)
     # plt.plot(dst_xb, dst_yb, 'k.')
     # plt.show()