… .travis.yml
… appveyor.yml
… pyresample/gradient/_gradient_search.c
… pyresample/gradient/_gradient_search.pyx
… pyresample/test/test_gradient.py
… setup.cfg
… setup.py