This PR adds filtering to gradient search resampling that discards source data chunks that do not intersect the target area.

The gradient search is now an optional mode, and requires `shapely` to be installed.

With this PR the following work with gradient search resampling:

- resampling of datasets with different resolutions
- resampling of datasets that previously failed with `IndexError: Bad index` because chunks had incompatible shapes
- resampling of data with `AreaDefinition`
- resampling of data with `SwathDefinition` with "smooth" coordinates (= not VIIRS, MODIS, point data)

- [x] Tests added
- [x] Tests passed
- [x] Passes ``git diff origin/master **/*py | flake8 --diff``