@pnuu 2020-05-14
Some timing results so far for the following script:

```python
import glob
import os
os.environ['PYTROLL_CHUNK_SIZE'] = '1024'
os.environ['DASK_NUM_WORKERS'] = '1'
os.environ['OMP_NUM_THREADS'] = '1'

from satpy import Scene

fnames = glob.glob('/home/lahtinep/data/satellite/geo/msg/*202004170600*')
glbl = Scene(reader='seviri_l1b_hrit', filenames=fnames)
glbl.load(['natural_color', 'overview', 'red_snow'], generate=False)
lcl = glbl.resample('eurol', resampler='gradient_search', reduce_data=False)
lcl.save_datasets(base_dir='/tmp')
```

- `master` branch: 31.1 s (1 core) / 23.1 s (2 cores) / 20.6 s (4 cores)
- this pr: 23.6 s / 16.7 s / 14.7

This image shows how the data was reduced before resampling (black = target area, blue = chunks used in processing, red = discarded chunks):
![geos_crop_boundaries_fixed_covers_seviri](https://user-images.githubusercontent.com/3170788/81932790-fd276e00-95f4-11ea-82bb-5b2cbb406182.png)

There are still some problems with ABI data. Although the chunk filtering works ...
![geos_crop_boundaries_fixed_covers](https://user-images.githubusercontent.com/3170788/81932933-3364ed80-95f5-11ea-8623-c1ef93a1bee1.png)

... the image still doesn't have full coverage:
![airmass_20170727_090038](https://user-images.githubusercontent.com/3170788/81933015-542d4300-95f5-11ea-95bf-68a157322646.png)