@pnuu 2020-05-14
The problem with ABI was my assumption of how the chunks are reordered for processing. The stepping goes first in Y direction, then in X. I've got it reversed. It works wit Seviri because there are only one chunk in X direction.