@pnuu 2020-05-14
The last commit fixed the chunk collection for ABI, so here are some timing results for ABI L1b (reduced resolution via Eumetcast) for three RGBs (`['airmass', 'ash', 'night_fog']`), resampled to my test area `'conus_5km'` (area def shown below):

- `master` branch: 30.8 s (1 core) / 24.5 s (2 cores) / 22.5 s (4 cores)
- this pr: 11.7 s / 9.8 s / 9.0 s

The six chunks that were used in processing are shown in the above image in blue surrounding the black target area.

```yaml
conus_5km:
  description: Continental USA, 5 km resolution
  projection:
    proj: laea
    a: 6370997.0
    b: 6370997.0
    lat_0: 50.0
    lon_0: -100.0
  shape:
    height: 669
    width: 1027
  area_extent:
    lower_left_xy: [-2864809.2037627744, -2830762.316377244]
    upper_right_xy: [2268528.5715566175, 514757.94726551295]
```