@pnuu 2020-05-14
> So now after your last fix the ABI image is correct?

Yes:
![airmass_20170727_090038](https://user-images.githubusercontent.com/3170788/81938333-72973c80-95fd-11ea-9b31-b07a0ec19382.png)