@pnuu 2020-05-14
According to `dask.diagnostics.ResourceProfiler` the memory usage dropped from ~3.3 GB (master branch) to ~1.4 GB when using 4 Dask workers for the ABI case.