@pnuu 2020-05-15
With this latest commit, the timings are

- ABI: 11. 3 s (1 core), 9.4 s (2 cores), 11.3 s (4 cores)
- SEVIRI: 24.3 s (1 core). 16.6 (2 cores), 14.7 s (4 cores)

For ABI there's a slight improvement even with this rather small target area, and with larger target area (= more chunks) the gain would be larger. There's zero gain for SEVIRI as all the used source chunks cover all the target chunks.