@pnuu 2020-05-15
Another test with 1 km CONUS area (3345 x 5135) and ABI:

- `master`: 5 m 11.5 s
- filter only source chunks: 1 m 7.3 s
- filter also target chunks (the latest commit): 1 m 9.4 s
- filter also target chunks, cache polygons: 1 m 7.5 s

Pretty significant jump from the `master` branch, I'd say.

To get the most of the target chunk filtering, we'd need to pre-compute all the boundary polygons, but I didn't find a way to pass them via `da.blockwise()` to the worker function. Or if there were even a way to pass some info on the chunk location it'd be possible to access them from each of the worker.