@pnuu 2020-05-19
After the last major restructuring (pair-up the input and output chunks for processing) the 1 km CONUS case takes 1 m 13 s to process and uses 10 GB of memory. The previous iteration used only 2.5 GB, so next I need to look at reducing the memory usage.