@pnuu 2020-05-19
The memory consumption is now solved. I was doing a lot of unnecessary padding, stacking and `da.nanmax()`'ing. With the latest commit the ABI case takes 40.0 s to process with a memory use of 2.5 GB. Compared to the 5+ minutes and ~5 GB of the original version in `master` branch, I'm quite happy.