@pnuu 2020-05-22
Appveyor doesn't seem to be using conda to install the packages, so I don't know how to fix the linux build error.