@pnuu 2020-05-25
I copied some fixes from https://github.com/pytroll/pyresample/pull/247/ so that the arrays passed to Cython code are read-only (`const`) and makes distributed processing to work more reliably. This also required the C source code to be re-generated, which will probably cause a huge merge conflict.